ask-server/requirements.txt
requests
python-dotenv
Markdown
customtkinter
Pillow
pymupdf
python-docx
chromadb